[r0.1] - controllers typing change

+++ src/main/java/com/jr/api/math/SubtractionController.java
@@ -11,8 +11,8 @@ public class SubtractionController {
 
      //GET
     @GetMapping(value = "/{num0}/{num1}") 
-    public int subtraction(@PathVariable("num0") Long num0, @PathVariable("num1") Long num1){ 
-        return (int) (num0 - num1);
+    public Long subtraction(@PathVariable("num0") Long num0, @PathVariable("num1") Long num1){ 
+        return num0 - num1;
     }
     
 }
